==> Step1 AC and B analysis _ThailifeLeadDataPerparation.ipynb <==
# Setting

#importing some lib and set display table.
import numpy as np
import pandas as pd
pd.set_option('display.max_rows', 50)
pd.set_option('display.max_columns', None)

# the name of file meaning as below. 
# ac = account ,b = base, st = statement
# tr = Transaction ,ct = Contacting
ac_1218 = 'Customer_RL_Account_1218.txt'
ac_0119 = 'Customer_RL_Account_0119.txt'
ac_0219 = 'Customer_RL_Account_0219.txt'
b_1218 = 'Customer_RL_Base_1218.txt'
b_0119 = 'Customer_RL_Base_0119.txt'
b_0219 = 'Customer_RL_Base_0219.txt'
st_1218 = 'Customer_RL_Statement_1218.txt'
st_0119 = 'Customer_RL_Statement_0119.txt'
st_0219 = 'Customer_RL_Statement_0219.txt'
tr_1218 = 'Customer_RL_Transaction_1218.txt'
tr_0119 = 'Customer_RL_Transaction_0119.txt'
tr_0219 = 'Customer_RL_Transaction_0219.txt'
ct_1218 = 'Customer_RL_Contacting_1218.txt'
ct_0119 = 'Customer_RL_Contacting_0119.txt'
ct_0219 = 'Customer_RL_Contacting_0219.txt'

# 3 month Exploretion

# Account and Base
ac_1218_data = pd.read_csv(ac_1218,header = 0,sep = '|',encoding="utf-8-sig")
ac_0119_data = pd.read_csv(ac_0119,header = 0,sep = '|',encoding="utf-8-sig")
ac_0219_data = pd.read_csv(ac_0219,header = 0,sep = '|',encoding="utf-8-sig")
b_1218_data = pd.read_csv(b_1218,header = 0,sep = '|',encoding="ISO-8859-1")
b_0119_data = pd.read_csv(b_0119,header = 0,sep = '|',encoding="ISO-8859-1")
b_0219_data = pd.read_csv(b_0219,header = 0,sep = '|',encoding="ISO-8859-1")

# Casting type and fill&drop NAN data
import datetime as dt
now = dt.datetime.now()

AllDataAC = [ac_1218_data,ac_0119_data,ac_0219_data]
for d in AllDataAC:
    d['CUSTOMER_ID'] = d['CUSTOMER_ID'].astype('str')
    d['Account_Number'] = d['Account_Number'].astype('str')
    d['Account_Status'] = d['Account_Status'].astype('str')
    d['Account_Type'] = d['Account_Type'].astype('str')
    d['Open_Date'] = d['Open_Date'].astype('str')
    d['Card_Age'] = d['Open_Date'].str.slice(0, 4)
    d['Credit_Limit'] = d['Credit_Limit'].astype('float')

# calculate Card_Age
AllDataAC = [ac_1218_data,ac_0119_data,ac_0219_data]
for d in AllDataAC:
    d['Card_Age'] = d['Card_Age'].astype('int')

AllDataAC = [ac_1218_data,ac_0119_data,ac_0219_data]
for d in AllDataAC:
    d['Card_Age'] = now.year - d['Card_Age']

AllDataB = [b_1218_data,b_0119_data,b_0219_data]
for d in AllDataB:
    d['CUSTOMER_ID'] = d['CUSTOMER_ID'].astype('str')
    d['DOB'] = d['DOB'].astype('str')
    d['Age'] = d['DOB'].str.slice(0, 4)
    d['Gender'] = d['Gender'].astype('str')
    d['Marital_status'] = d['Marital_status'].fillna("Unknown")
    d['Marital_status'] = d['Marital_status'].astype('str')
    d['Mobilephone_fg'] = d['Mobilephone_fg'].astype('str')
    d['Officephone_fg'] = d['Officephone_fg'].astype('str')
    d['Occupation'] = d['Occupation'].fillna("Unknown")
    d['Occupation'] = d['Occupation'].replace({'':'Unknown',' ':'Unknown'})
    d['Occupation'] = d['Occupation'].astype('str')
    d['Province'] = d['Province'].astype('str')
    d['Zipcode'] = d['Zipcode'].astype('str')
    d['Customer_Status'] = d['Customer_Status'].astype('str')
    d['Is_Staff'] = d['Is_Staff'].astype('str')
    d['Is_KTC_Credit_Card_Holder'] = d['Is_KTC_Credit_Card_Holder'].astype('str')
    d = d.drop(columns=['education','home_stat'])

# calculate Age
AllDataB = [b_1218_data,b_0119_data,b_0219_data]
for d in AllDataB:
    d['Age'] = d['Age'].astype('int')

AllDataB = [b_1218_data,b_0119_data,b_0219_data]
for d in AllDataB:
    d['Age'] = now.year - d['Age']
    

#Drop usedless collumns
ac_1218_data = ac_1218_data.drop(columns=['Close_Date'])
ac_0119_data = ac_0119_data.drop(columns=['Close_Date'])
ac_0219_data = ac_0219_data.drop(columns=['Close_Date'])
b_1218_data = b_1218_data.drop(columns=['education','home_stat'])
b_0119_data = b_0119_data.drop(columns=['education','home_stat'])
b_0219_data = b_0219_data.drop(columns=['education','home_stat'])

#merge
b_ac_1218_data = b_1218_data.merge(ac_1218_data, left_on=['CUSTOMER_ID'], right_on=['CUSTOMER_ID'], how='inner')
b_ac_0119_data = b_0119_data.merge(ac_0119_data, left_on=['CUSTOMER_ID'], right_on=['CUSTOMER_ID'], how='inner')
b_ac_0219_data = b_0219_data.merge(ac_0219_data, left_on=['CUSTOMER_ID'], right_on=['CUSTOMER_ID'], how='inner')

# save to csv
b_ac_1218_data.to_csv("b_ac_1218_data.csv",encoding="utf-8-sig")
b_ac_0119_data.to_csv("b_ac_0119_data.csv", encoding="utf-8-sig")
b_ac_0219_data.to_csv("b_ac_0219_data.csv", encoding="utf-8-sig")
