==> Step5 Contaction.ipynb <==
# Setting

## Create file from contact and prepare training file.

#importing some lib and set display table.
import numpy as np
import pandas as pd
pd.set_option('display.max_rows', 50)
pd.set_option('display.max_columns', None)

# the name of file meaning as below. 
# ac = account ,b = base, st = statement
# tr = Transaction ,ct = Contacting
# b_ac = b join with ac by using 3 AC and B analysis file.
# b_ac_st_t = b_ac_st join with t by using  T analysis file.
ac_1218 = 'Customer_RL_Account_1218.txt'
ac_0119 = 'Customer_RL_Account_0119.txt'
ac_0219 = 'Customer_RL_Account_0219.txt'
b_1218 = 'Customer_RL_Base_1218.txt'
b_0119 = 'Customer_RL_Base_0119.txt'
b_0219 = 'Customer_RL_Base_0219.txt'
st_1218 = 'Customer_RL_Statement_1218.txt'
st_0119 = 'Customer_RL_Statement_0119.txt'
st_0219 = 'Customer_RL_Statement_0219.txt'
tr_1218 = 'Customer_RL_Transaction_1218.txt'
tr_0119 = 'Customer_RL_Transaction_0119.txt'
tr_0219 = 'Customer_RL_Transaction_0219.txt'
ct_1218 = 'Customer_RL_Contacting_1218.txt'
ct_0119 = 'Customer_RL_Contacting_0119.txt'
ct_0219 = 'Customer_RL_Contacting_0219.txt'
b_ac_1218 = 'b_ac_1218_data.csv'
b_ac_0119 = 'b_ac_0119_data.csv'
b_ac_0219 = 'b_ac_0119_data.csv'
b_ac_st_1218 = 'b_ac_st_1218_data.csv'
b_ac_st_0119 = 'b_ac_st_0119_data.csv'
b_ac_st_0219 = 'b_ac_st_0219_data.csv'
b_ac_st_t_1218 = 'b_ac_st_t_1218_data.csv'
b_ac_st_t_0119 = 'b_ac_st_t_0119_data.csv'
b_ac_st_t_0219 = 'b_ac_st_t_0219_data.csv'

ct_data = pd.read_csv(ct_0219 ,header = 0,sep = '|', encoding="ISO-8859-1")

ct_data = ct_data[ct_data["Curr_Cpgn_Insurer"] == "TCL"]
ct_data = ct_data[(ct_data['Curr_Cpgn_Product_Name'] =="New Perfect Protection")]
ct_data = ct_data[['CUSTOMER_NO','Curr_Cpgn_Last_Call_Date','Curr_Cpgn_Response_Flag']]
ct_data['CUSTOMER_NO'] = ct_data['CUSTOMER_NO'].astype('str')
ct_data['Curr_Cpgn_Last_Call_Date'] = ct_data['Curr_Cpgn_Last_Call_Date'].astype('str')
ct_data = ct_data[ct_data['Curr_Cpgn_Last_Call_Date'].str.contains("201901")]# chance date slice


bese_data = pd.read_csv('summarize_b_ac_st_t_0219.csv', encoding = "utf-8-sig", index_col = 0)

bese_data['CUSTOMER_ID'] = bese_data['CUSTOMER_ID'].astype('str')
Train_data=bese_data.merge(ct_data, left_on=['CUSTOMER_ID'], right_on=['CUSTOMER_NO'], how ='inner').sort_values(['CUSTOMER_ID'])

col = ['CUSTOMER_ID','Gender','MONTHLY_INCOME_RL','Marital_status','Mobilephone_fg','Officephone_fg',\
       'Homephone_fg','NO_OF_CHILD','Occupation','Total_Credit_Limit' ,'Customer_Status',\
       'Is_Staff','Is_KTC_Credit_Card_Holder' ,'Age','Credit_Limit', 'Current_Balance','Card_Age', \
      'CyC_Balance_Size','CyC_Balance_Min_Score','Cyc_Balance_PT_Max_Score','CyC_Balance_Mean','Cyc_Balance_Std',\
       'Cyc_Balance_PT_5_Score','Cyc_Balance_PT_25_Score','Cyc_Balance_PT_50_Score','Cyc_Balance_PT_75_Score','Cyc_Balance_PT_95_Score',\
       'Payment_Size','Payment_Mean','Payment_Std','Payment_Min_Score','Payment_PT_5_Score','Payment_PT_25_Score','Payment_PT_50_Score',\
       'Payment_PT_75_Score','Payment_PT_95_Score','Payment_Max_Score','Payment_sum',\
       'Cash_Size','Cash_Mean','Cash_Std','Cash_Min_Score','Cash_PT_5_Score','Cash_PT_25_Score',\
       'Cash_PT_50_Score','Cash_PT_75_Score','Cash_PT_95_Score','Cash_Max_Score','Cash_sum','Curr_Cpgn_Response_Flag']

Train_data = Train_data[col] 
Train_data = Train_data.dropna(axis = 0)

Train_data.to_csv('training_b_ac_st_t_0219.csv', encoding = "utf-8-sig")

Train_data.info()

ct_data.info()

Train_data[Train_data['Curr_Cpgn_Response_Flag'] == 'Y'].count()

Train_data[Train_data['Curr_Cpgn_Response_Flag'] == 'Y'].describe()
