==> Step6.1 AutoML(Tpot).ipynb <==
#importing some lib and set display table.
import numpy as np
import pandas as pd
pd.set_option('display.max_rows', 50)
pd.set_option('display.max_columns', None)

Feature_col  = ['Gender','MONTHLY_INCOME_RL','Marital_status','Mobilephone_fg','Officephone_fg',\
       'Homephone_fg','NO_OF_CHILD','Occupation','Total_Credit_Limit',\
       'Is_Staff','Is_KTC_Credit_Card_Holder' ,'Age','Credit_Limit', 'Current_Balance','Card_Age', \
      'CyC_Balance_Size','CyC_Balance_Min_Score','Cyc_Balance_PT_Max_Score','CyC_Balance_Mean','Cyc_Balance_Std',\
       'Cyc_Balance_PT_5_Score','Cyc_Balance_PT_25_Score','Cyc_Balance_PT_50_Score','Cyc_Balance_PT_75_Score','Cyc_Balance_PT_95_Score',\
       'Payment_Size','Payment_Mean','Payment_Std','Payment_Min_Score','Payment_PT_5_Score','Payment_PT_25_Score','Payment_PT_50_Score',\
       'Payment_PT_75_Score','Payment_PT_95_Score','Payment_Max_Score','Payment_sum',\
       'Cash_Size','Cash_Mean','Cash_Std','Cash_Min_Score','Cash_PT_5_Score','Cash_PT_25_Score',\
       'Cash_PT_50_Score','Cash_PT_75_Score','Cash_PT_95_Score','Cash_Max_Score','Cash_sum']

Target_col = ['Curr_Cpgn_Response_Flag']
t_data = 'training_0119_to_0219.csv'

all_data = 'summarize_b_ac_st_t_0219.csv'

df = pd.read_csv(t_data, encoding = 'utf-8-sig', index_col = 0)
df = df.drop(columns=['Customer_Status'])

#Cat code
YN_types = ['Y', 'N'] #yes/no
FM_types = ['Female','Male'] #Gender
M_types = ['Single','Married','Unknown']
Oc_types = ['55', 'Not Answer', 'Permanent Employee', 'Doctor / Dentist',
       'Business Owner - Company Registration', 'Lawyer / Judiciary',
       'Auditor', 'Other', 'Soldier', 'Small Constructor', 'Merchant',
       'Marketing Ofiicer', 'Financial Officer',
       'Business Owner - Juristic Person', 'Chief Accountant', 'Salesman',
       'KTC Employee', 'Engineer', 'Agriculturist', 'Politician',
       'Teacher in University', 'Pilot', 'Attroney',
       'Air Hostess / Steward', 'Teacher in Up country', 'Tecniciant',
       'Teacher', 'Architect', 'Policeman', 'Not Ask', 'Secretary',
       'Phamacist', 'Nurse', 'Survey Officer / Map Officer', 'Programmer',
       'Librarian', 'Purchasing Officer',
       'Journalist / Editor / Writer/ Novelist', 'Guide', 'Driver',
       'Export Officer', 'Designer', 'Worker', 'Taxi Driver', 'Masseur',
       'Security', 'Scientist', '60', 'Tailor / Designer',
       'Maid / Janitor', 'Fisherman', 'Investment Advisor', 'Fireman',
       'Messenger', 'Barber', 'CV', 'Delivery Officer',
       'Artist / A Star / Singer / Sportman', 'Economist', '66',
       'Prison Controller', 'Diplomat', '61', '72', 'Pupil / Student',
       '65', 'Waiter / Waitress'] # Occupation

df['Gender'] = df.Gender.astype("category", categories=FM_types).cat.codes
df['Marital_status'] = df.Marital_status.astype("category", categories=M_types).cat.codes
df['Mobilephone_fg'] = df.Mobilephone_fg.astype("category", categories=YN_types).cat.codes
df['Officephone_fg'] = df.Officephone_fg.astype("category", categories=YN_types).cat.codes
df['Homephone_fg'] = df.Homephone_fg.astype("category", categories=YN_types).cat.codes
df['Occupation'] = df.Occupation.astype("category", categories=Oc_types).cat.codes
df['Is_Staff'] = df.Is_Staff.astype("category", categories=YN_types).cat.codes
df['Is_KTC_Credit_Card_Holder'] = df.Is_KTC_Credit_Card_Holder.astype("category", categories=YN_types).cat.codes
df['Curr_Cpgn_Response_Flag'] = df.Curr_Cpgn_Response_Flag.astype("category", categories=YN_types).cat.codes

from sklearn.model_selection import train_test_split
X = df[Feature_col]
Y = df[Target_col]
X_train, X_test, Y_train, Y_test = train_test_split(X, Y, test_size=0.4)

from tpot import TPOTClassifier
tpot = TPOTClassifier(generations=5, population_size=50, verbosity=2)
tpot.fit(X_train, Y_train)
print("train score::", tpot.score(X_train, Y_train))
print("test score::", tpot.score(X_test, Y_test))
tpot.export('tpot_KTC_PL23_pipeline.py')

Y_predict_train = tpot.predict(X_train)
Y_predict_test = tpot.predict(X_test)

from sklearn.metrics import confusion_matrix
print("The Confusion Metrix of feature of train set")
unique_label = np.unique(Y_train)
print(pd.DataFrame(confusion_matrix(Y_train,
                                    Y_predict_train,labels=unique_label), 
                index=['true:{:}'.format(x) for x in unique_label], 
                columns=['pred:{:}'.format(x) for x in unique_label]))

from sklearn.metrics import confusion_matrix
print("The Confusion Metrix of feature of test set")
unique_label = np.unique(Y_test)
print(pd.DataFrame(confusion_matrix(Y_test,
                                    Y_predict_test,labels=unique_label), 
                index=['true:{:}'.format(x) for x in unique_label], 
                columns=['pred:{:}'.format(x) for x in unique_label]))
