==> Step6.2 PermutationImpotant.ipynb <==
#importing some lib and set display table.
import numpy as np
import pandas as pd
pd.set_option('display.max_rows', 50)
pd.set_option('display.max_columns', None)

Feature_col  = ['Gender','MONTHLY_INCOME_RL','Marital_status','Mobilephone_fg','Officephone_fg',\
       'Homephone_fg','NO_OF_CHILD','Occupation','Total_Credit_Limit',\
       'Is_Staff','Is_KTC_Credit_Card_Holder' ,'Age','Credit_Limit', 'Current_Balance','Card_Age', \
      'CyC_Balance_Size','CyC_Balance_Min_Score','Cyc_Balance_PT_Max_Score','CyC_Balance_Mean','Cyc_Balance_Std',\
       'Cyc_Balance_PT_5_Score','Cyc_Balance_PT_25_Score','Cyc_Balance_PT_50_Score','Cyc_Balance_PT_75_Score','Cyc_Balance_PT_95_Score',\
       'Payment_Size','Payment_Mean','Payment_Std','Payment_Min_Score','Payment_PT_5_Score','Payment_PT_25_Score','Payment_PT_50_Score',\
       'Payment_PT_75_Score','Payment_PT_95_Score','Payment_Max_Score','Payment_sum',\
       'Cash_Size','Cash_Mean','Cash_Std','Cash_Min_Score','Cash_PT_5_Score','Cash_PT_25_Score',\
       'Cash_PT_50_Score','Cash_PT_75_Score','Cash_PT_95_Score','Cash_Max_Score','Cash_sum']

Target_col = ['Prediction']

all_data = 'summarize_predict_b_ac_st_t_0219.csv'

from sklearn.externals import joblib
filename = 'finalized_model.sav'
# some time later...
 
# load the model from disk
loaded_model = joblib.load(filename)

#Load all data.
df_all = pd.read_csv(all_data, encoding = 'utf-8-sig', index_col = 0)
df_all = df_all.dropna(axis = 0)
df_for_predict = df_all.copy()

#Cat code
YN_types = ['Y', 'N'] #yes/no
FM_types = ['Female','Male'] #Gender
M_types = ['Single','Married','Unknown']
Oc_types = df_all.Occupation.unique() # Occupation

df_for_predict['Gender'] = df_for_predict.Gender.astype("category", categories=FM_types).cat.codes
df_for_predict['Marital_status'] = df_for_predict.Marital_status.astype("category", categories=M_types).cat.codes
df_for_predict['Mobilephone_fg'] = df_for_predict.Mobilephone_fg.astype("category", categories=YN_types).cat.codes
df_for_predict['Officephone_fg'] = df_for_predict.Officephone_fg.astype("category", categories=YN_types).cat.codes
df_for_predict['Homephone_fg'] = df_for_predict.Homephone_fg.astype("category", categories=YN_types).cat.codes
df_for_predict['Occupation'] = df_for_predict.Occupation.astype("category", categories=Oc_types).cat.codes
df_for_predict['Is_Staff'] = df_for_predict.Is_Staff.astype("category", categories=YN_types).cat.codes
df_for_predict['Is_KTC_Credit_Card_Holder'] = df_for_predict.Is_KTC_Credit_Card_Holder.astype("category", categories=YN_types).cat.codes

new_features =  ['Gender','MONTHLY_INCOME_RL','Officephone_fg','Homephone_fg','NO_OF_CHILD','Occupation','Age','Credit_Limit',
  'Current_Balance','Card_Age','CyC_Balance_Size','CyC_Balance_Min_Score','Cyc_Balance_PT_Max_Score',
  'CyC_Balance_Mean','Cyc_Balance_Std','Cyc_Balance_PT_5_Score','Cyc_Balance_PT_25_Score','Cyc_Balance_PT_75_Score',
  'Cyc_Balance_PT_95_Score','Payment_Mean','Payment_Std','Payment_Min_Score','Payment_PT_5_Score','Payment_PT_25_Score',
  'Payment_PT_50_Score','Payment_PT_75_Score','Payment_PT_95_Score','Payment_Max_Score','Payment_sum',
  'Cash_Size','Cash_Mean','Cash_Std','Cash_Min_Score','Cash_PT_5_Score','Cash_PT_25_Score','Cash_PT_50_Score','Cash_PT_75_Score',
  'Cash_PT_95_Score','Cash_Max_Score','Cash_sum']


import eli5
from eli5.sklearn import PermutationImportance
val_X = df_for_predict[new_features]
val_Y = df_for_predict[Target_col]
perm = PermutationImportance(loaded_model, random_state=1).fit(val_X, val_Y)
eli5.show_weights(perm, feature_names = val_X.columns.tolist())
